+++ packages/cdktf/lib/app.ts
@@ -1,9 +1,9 @@
 import { Construct, Node, ConstructMetadata } from 'constructs';
 import * as fs from 'fs';
 import { version } from '../package.json';
+import { Manifest } from './manifest'
 
 export const CONTEXT_ENV = 'CDKTF_CONTEXT_JSON';
-
 export interface AppOptions {
     /**
      * The directory to output Terraform resources.
@@ -35,12 +35,18 @@ export class App extends Construct {
     public readonly outdir: string;
 
     /**
+     * The stack which will be synthesized. If not set, all stacks will be synthesized.
+     */
+    public readonly targetStackId: string | undefined
+
+    /**
      * Defines an app
      * @param options configuration options
      */
     constructor(options: AppOptions = {}) {
         super(undefined as any, '');
         this.outdir = process.env.CDKTF_OUTDIR ?? options.outdir ?? 'cdktf.out';
+        this.targetStackId = process.env.CDKTF_TARGET_STACK_ID
 
         this.loadContext(options.context);
 
@@ -56,13 +62,20 @@ export class App extends Construct {
      * Synthesizes all resources to the output directory
      */
     public synth(): void {
-        if (!fs.existsSync(this.outdir)) {
-            fs.mkdirSync(this.outdir);
-        }
+      if (!fs.existsSync(this.outdir)) {
+        fs.mkdirSync(this.outdir);
+      }
 
-        Node.of(this).synthesize({
-            outdir: this.outdir
-        });
+      const manifest = new Manifest(version, this.outdir)
+
+      Node.of(this).synthesize({
+        outdir: this.outdir,
+        sessionContext: {
+          manifest
+        }
+      });
+
+      manifest.writeToFile();
     }
 
     private loadContext(defaults: { [key: string]: string } = { }) {
@@ -78,7 +91,7 @@ export class App extends Construct {
         const contextFromEnvironment = contextJson
           ? JSON.parse(contextJson)
           : { };
-    
+
         for (const [k, v] of Object.entries(contextFromEnvironment)) {
           node.setContext(k, v);
         }