+++ test/jest.config.js
@@ -4,10 +4,12 @@ module.exports = {
   roots: [
     "<rootDir>"
   ],
+  setupFilesAfterEnv: ['./jest.setup.js'],
   testMatch: ['**/*test.ts'],
   moduleFileExtensions: [
     "js",
     "ts"
   ],
-  runner: "groups"
+  runner: "groups",
+  testTimeout: 300000
 }