+++ test/typescript/feature-flags/test.ts
@@ -18,39 +18,39 @@ describe("full integration test", () => {
     ]
   }
 
-  beforeAll(() => {
+  beforeAll(async () => {
     driver = new TestDriver(__dirname)
     driver.switchToTempDir()
-    driver.init('typescript')
+    await driver.init('typescript')
     driver.copyFiles('main.ts')
     writeConfig(driver.workingDirectory, cdktfJSON)
-    driver.get()
+    await driver.get()
   });
 
-  test("with excludeStackIdFromLogicalIds feature", () => {
+  test("with excludeStackIdFromLogicalIds feature", async () => {
     writeConfig(driver.workingDirectory, jsonWithContext({ excludeStackIdFromLogicalIds: "true" }))
-    driver.synth()
-    expect(loadStackJson(driver.workingDirectory)).toMatchSnapshot();
+    await driver.synth()
+    expect(loadStackJson(driver.workingDirectory, 'hello-deploy')).toMatchSnapshot();
   });
 
-  test("with allowSepCharsInLogicalIds feature", () => {
+  test("with allowSepCharsInLogicalIds feature", async () => {
     writeConfig(driver.workingDirectory, jsonWithContext({ allowSepCharsInLogicalIds: "true" }))
-    driver.synth()
-    expect(loadStackJson(driver.workingDirectory)).toMatchSnapshot();
+    await driver.synth()
+    expect(loadStackJson(driver.workingDirectory, 'hello-deploy')).toMatchSnapshot();
   });
 
-  test("without features", () => {
+  test("without features", async () => {
     writeConfig(driver.workingDirectory, cdktfJSON)
-    driver.synth()
-    expect(loadStackJson(driver.workingDirectory)).toMatchSnapshot();
+    await driver.synth()
+    expect(loadStackJson(driver.workingDirectory, 'hello-deploy')).toMatchSnapshot();
   });
 
   const jsonWithContext = (context) => {
     return Object.assign({}, cdktfJSON, { context })
   }
 
-  const loadStackJson = (workingDir) => {
-    return fs.readFileSync(path.join(workingDir, 'cdktf.out', 'cdk.tf.json'), 'utf-8')
+  const loadStackJson = (workingDir, stackName) => {
+    return fs.readFileSync(path.join(workingDir, 'cdktf.out', 'stacks', stackName, 'cdk.tf.json'), 'utf-8')
   }
 
   const writeConfig = (workingDir, json) => {