+++ test/typescript/synth-app/test.ts
@@ -9,13 +9,13 @@ import { TestDriver } from "../../test-helper";
 describe("full integration test synth", () => {
   let driver: TestDriver;
 
-  beforeAll(() => {
+  beforeAll(async () => {
     driver = new TestDriver(__dirname)
-    driver.setupTypescriptProject()
+    await driver.setupTypescriptProject()
   });
 
   test("synth generates JSON", async () => {
-    driver.synth()
-    expect(driver.synthesizedStack()).toMatchSnapshot()
+    await driver.synth()
+    expect(driver.synthesizedStack('hello-terra')).toMatchSnapshot()
   })
 })